Fix exceptions thrown by ShadowSQLiteConnection
In addition to making the exceptions thrown match those thrown from
android, this CL improves the stack traces - all exceptions thrown
include the ExecutionException in them, which shows the real invocation
site of the command throwing the exception, not just the stack trace of
ShadowSQLiteConnection’s internal executor.

Also, interrupting the thread making a call to the database no longer
results in an exception.